Tease Encryption apart into handlers
